fixed multiply by zero
